power_profile: Add cpu.awake parameter
One of the remaining issues with the new profile is lack
of a cpu.awake parameter - without this, wakelock holders
won't be penalized in battery stats.

Make a wild-ass guesstimate of 15 mA, based on:
Best-case consumption is around 0.5%/hour with radio on, assume
8.25 mA consumption of non-CPU stuff in deep sleep.
Best case consumption when wakelocked is usually around 1.25-1.5%/hour,
assume then, about 25 mA total when wakelocked.
Subtract the baseline and it's 15-16 mA.  Let's just round to a
multiple of 5 here, since it IS a WAG.

Also, consumption when the CPU is running at 200 MHz (for example,
governor set to ignore nice load, and a program is endlessly looping
but niced to avoid bumping the gov) is usually around 4-5%/hour,
which is semi-consistent with a number of 55 mA for 200 MHz op.

Change-Id: I6aa08842dd2acfced0c70a95d973aa0b415799cc